% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_error.R
\name{standard_error}
\alias{standard_error}
\title{Standard Error Fundtion}
\usage{
standard_error(v)
}
\arguments{
\item{v}{a numeric vector}
}
\value{
standard error
}
\description{
Standard Error Fundtion
}
\examples{
standard_error()
}
