% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_squares.R
\name{sum_squares}
\alias{sum_squares}
\title{Sum Squares Function}
\usage{
sum_squares(v)
}
\arguments{
\item{v}{is a vector of numbers to use to compute sum squares}
}
\value{
number
}
\description{
Sum Squares Function
}
\examples{
height <- rnorm(100, mean = 10, sd = 0.8)
sum_squares(height)
}
